Added in gambit button, and name display functionality

+++ Assets/Scripts/Battle/BattleStats.cs
@@ -38,7 +38,7 @@ public class BattleStats
     public int BattlerNumber { get; private set; }
 
     public int BattlerCurrentHp { get; private set; }
-    public int BattlerCurrentHpPercent => BattlerCurrentHp / BattlerMaxHp * 100;
+    public int BattlerCurrentHpPercent() => (int)((float)BattlerCurrentHp / BattlerMaxHp * 100);
     public int BattlerCurrentMpPercent => BattlerCurrentMp / BattlerMaxMp * 100;
     public int BattlerCurrentMp { get; private set; }
     public int BattlerMaxMp => _theBattlersBaseBaseStats.BattlerMp;
